+++ bindings/go/setup_test.go
@@ -1,4 +1,4 @@
-package ariaengine
+package aria
 
 import (
 	"os"