+++ bindings/flutter/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 0.1.0
 
-- Initial Flutter FFI binding (`import 'package:ariaengine/ariaengine.dart'`).
+- Initial Flutter FFI binding (`import 'package:aria_engine/aria_engine.dart'`).